BIG changes
Couple of real challenges the past few days. After loading the board i decided to start implementing the tutorial and running the user through piece functionality. The game doesn't have any logic and relies on updating the Pixmap of the pieces new position and clearing the pixmap of the pieces previous position. This is something i will have to check with a project coordinator.